+++ CHANGELOG.md
@@ -1,14 +1,26 @@
 # Change Log
 All notable changes to QingStor SDK for Swift will be documented in this file.
 
+## [v2.1.0] - 2016-12-26
+
+### Changed
+- Fix signer bug.
+- Add more parameters to sign.
+
+### Added
+- Add request parameters for GET Object.
+- Add IP address conditions for bucket policy.
+
 ## [v2.0.1] - 2016-12-15
+
 ### Changed
 - Improve the implementation of deleting multiple objects.
 
 ## [v2.0.0] - 2016-12-14
+
 ### Added
 - QingStor SDK for the Swift programming language.
 
-[Unreleased]: https://github.com/yunify/qingstor-sdk-swift/compare/v2.0.1...HEAD
+[v2.1.0]: https://github.com/yunify/qingstor-sdk-swift/compare/v2.0.1...v2.1.0
 [v2.0.1]: https://github.com/yunify/qingstor-sdk-swift/compare/v2.0.0...v2.0.1
 [v2.0.0]: https://github.com/yunify/qingstor-sdk-swift/compare/v2.0.0...v2.0.0